+++ src/plugins/core/widgets/links/LinkInput.tsx
@@ -1,10 +1,12 @@
 import * as React from 'react';
-import { IconButton, removeIcon } from '../../../../app/ui';
+import { arrowUpIcon, arrowDownIcon, IconButton, removeIcon } from '../../../../app/ui';
 import { Link as LinkProps } from './interfaces';
 
 interface Props extends LinkProps {
   number: number;
   onChange: (values: Partial<LinkProps>) => void;
+  onMoveUp?: () => void;
+  onMoveDown?: () => void;
   onRemove: () => void;
 }
 
@@ -13,6 +15,12 @@ const LinkInput: React.StatelessComponent<Props> = (props) => (
     <h5>
       <div className="title--buttons">
         <IconButton onClick={props.onRemove} title="Remove link">{removeIcon}</IconButton>
+        {props.onMoveDown !== undefined &&
+          <IconButton onClick={props.onMoveDown} title="Move link down">{arrowDownIcon}</IconButton>
+        }
+        {props.onMoveUp !== undefined &&
+          <IconButton onClick={props.onMoveUp} title="Move link up">{arrowUpIcon}</IconButton>
+        }
       </div>
 
       {props.number <= 9 ? `Keyboard shortcut ${props.number}` : 'Shortcut'}