+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - Simplified Chinese translations!
+- Ability to reorder Quick Links
 
 ## [1.10.1] - 2018-02-10
 ### Fixed